fix: 미션에서 주어진 프론트 코드 추가 및 버그 수정
BinaryContent를 스토리지에 저장할때, ReadStatus의 생성 등 에서 BinaryContent의 id, Channel의 id, createdAt등의 변수는 repository에 저장할 때 자동 생성되는데, 순서를 지키지 않아 해당 값들이 null로 들어가던 문제 수정

BasicChannelService의 findAllByUserId 로직이 PUBLIC 채널을 조회하지 않던 문제 수정

ReadStatusRepository의 일부 메서드명 변경

BasicUserService의 update 메서드에서 유저가 프로필을 가지고 있지 않을 때의 null 체크 로직 추가

MessageController의 findAllByChannelId 메서드에 페이지네이션 추가

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -45,8 +45,9 @@ public class BasicUserService implements UserService {
               dto.fileName(),
               dto.contentType(),
               dto.bytes().length);
+          binaryContentRepository.save(binaryContent);
           binaryContentStorage.put(binaryContent.getId(), dto.bytes());
-          return binaryContentRepository.save(binaryContent);
+          return binaryContent;
         })
         .findFirst()
         .orElse(null);
@@ -92,13 +93,14 @@ public class BasicUserService implements UserService {
               dto.fileName(),
               dto.contentType(),
               dto.bytes().length);
+          binaryContentRepository.save(binaryContent);
           binaryContentStorage.put(binaryContent.getId(), dto.bytes());
-          return binaryContentRepository.save(binaryContent);
+          return binaryContent;
         })
         .findFirst()
         .orElse(null);
 
-    if (newProfile != null) {
+    if (newProfile != null && user.getProfile() != null) {
       binaryContentRepository.deleteById(user.getProfile().getId());
     }
